added the separate file for SQL commands
