+++ configuration/config.py
@@ -1,4 +1,5 @@
 import user.operations as op
+import user.election_operations as ep
 
 input_line = "Enter your {}....:"
 
@@ -15,14 +16,14 @@ roles = {
 }
 role_function_mapping = {
     "Edit_details": op.AllOperation.edit_details,
-    "Give_vote": op.AllOperation.give_vote,
-    "Check_on_going_Elections": op.AllOperation.check_on_going_elections,
-    "Start_Elections": op.AllOperation.start_election,
-    "Add_Party": op.AllOperation.add_party,
+    "Give_vote": ep.AllOperation.give_vote,
+    "Check_on_going_Elections": ep.AllOperation.check_on_going_elections,
+    "Start_Elections": ep.AllOperation.start_election,
+    "Add_Party": ep.AllOperation.add_party,
     "Check_if_user_is_approved": op.AllOperation.is_approved,
     "Make_admin": op.AllOperation.make_admin,
-    "Close_Elections": op.AllOperation.close_elections,
-    "Election_Results": op.AllOperation.results,
+    "Close_Elections": ep.AllOperation.close_elections,
+    "Election_Results": ep.AllOperation.results,
     "Log_out": op.AllOperation.log_out,
     "Register_user": op.AllOperation.register_new_user,
     "Approve_the_user_for_login": op.AllOperation.approve_user_login,
